+++ R/as_workbook.R
@@ -4,7 +4,7 @@
 
 #' Convert a Tatoo Table Object to an Excel Workbook
 #'
-#' This function converts [`Tatoo_table`] or [`Tatoo_report`] objects directly
+#' `as_workbook()` converts [`Tatoo_table`] or [`Tatoo_report`] objects directly
 #' to [openxlsx] `Workbook` objects. For information about additional parameters
 #' please refer to the documentation of [write_worksheet()], for which
 #' `as_workbook()` is just a wrapper. Additional possible function arguments way
@@ -14,7 +14,8 @@
 #' @param ... Additional arguments passed on to `write_worksheet()`
 #'
 #' @family xlsx exporters
-#' @return an openxlsx openxlsx `Workbook` object (invisibly for `save_xlsx()`)
+#' @return `as_workbook()` returns an openxlsx `Workbook` object.
+#'
 #' @export
 #'
 #' @examples
@@ -41,7 +42,7 @@ as_workbook <- function(
   x,
   ...
 ){
-  assert_that(requireNamespace("openxlsx"))
+  require_openxlsx()
   UseMethod('as_workbook')
 }
 
@@ -107,6 +108,12 @@ as_workbook.Tatoo_report <- function(x, ...){
 
 
 
+#' @export
+as_workbook.Workbook <- function(x, ...){
+  x
+}
+
+
 # write_worksheet ---------------------------------------------------------
 
 #' Write Data to an openxlsx Worksheet
@@ -145,7 +152,7 @@ write_worksheet <- function(
   assert_that(is.scalar(sheet))
   assert_that(is.flag(append))
   assert_that(rlang::is_scalar_integerish(start_row))
-  assert_that(requireNamespace("openxlsx"))
+  require_openxlsx()
 
   UseMethod('write_worksheet')
 }
@@ -320,7 +327,7 @@ write_worksheet.Composite_table <- function(
 
   ## merge subtable heading cells
   for(i in seq_along(multinames)){
-    merge_start <- ifelse(i == 1L, 1, multinames[[i-1]] + 1)
+    merge_start <- ifelse(i == 1L, 1, multinames[[i-1]] + 1)  #nolint
     merge_end   <- multinames[[i]]
     openxlsx::mergeCells(
       wb,
@@ -474,13 +481,13 @@ write_worksheet.Stacked_table <- function(
 
 # save_xlsx ---------------------------------------------------------------
 
-#' `save_xlsx()` is a shortcut to save a `Tatoo_table` directly to local xlsx
-#' file.
+#' @description `save_xlsx()` is a wrapper for saving a `Tatoo_table` directly
+#'   to an \file{xlsx} file.
 #'
 #' @template outfile
 #' @template overwrite
 #'
-#' @return `TRUE` on success
+#' @return `save_xlsx()` returns the path to the saved \file{.xlsx} (invisibly).
 #' @export
 #' @rdname as_workbook
 #'
@@ -492,6 +499,7 @@ save_xlsx <- function(
 ){
   assert_that(rlang::is_scalar_character(outfile))
   assert_that(is.flag(overwrite))
+  require_openxlsx()
 
   UseMethod('save_xlsx')
 }
@@ -508,5 +516,24 @@ save_xlsx.default <- function(
 ){
   wb <- as_workbook(x, ...)
   openxlsx::saveWorkbook(wb, outfile, overwrite)
-  invisible(wb)
+  invisible(outfile)
+}
+
+
+
+
+#' @description  `view_xlsx()` is another wrapper for viewing a `Tatoo_table`'s
+#' \file{xlsx} representation in your favourite spreadsheet program (powered by
+#' [openxlsx::openXL()]).
+#'
+#' @return `view_xlsx()` opens an external program and returns `NULL` (invisibly).
+#'
+#' @rdname as_workbook
+#' @export
+view_xlsx <- function(
+  x,
+  ...
+){
+  openxlsx::openXL(as_workbook(x, ...))
+  invisible()
 }